+++ app/src/main/java/com/hadysalhab/movid/screen/main/featured/FeaturedFragment.kt
@@ -117,7 +117,7 @@ class FeaturedFragment : BaseFragment(), FeaturedView.Listener, FetchMovieGroups
         outState.putSerializable(SCREEN_STATE, screenState)
     }
 
-    override fun onMovieCardClicked(movieID: Long) {
+    override fun onMovieCardClicked(movieID: Int) {
         mainNavigator.toDetailFragment(movieID)
     }
